backend/requirements.txt
Flask
flask_sqlalchemy
flask_cors
PyMySQL
Werkzeug
python-dotenv

